requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
annotated-types==0.6.0
anyio==4.2.0
black==24.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fastapi==0.109.2
filelock==3.9.0
fonttools==4.48.1
h11==0.14.0
idna==3.4
isort==5.13.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
qiniu==7.13.0
requests==2.28.1
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
timm==0.6.7
torch==2.0.0+cpu
torchaudio==2.0.1+cpu
torchvision==0.15.1+cpu
typing_extensions==4.8.0
upload==0.1.1
urllib3==1.26.13
uvicorn==0.27.1
